+++ app/keepers/keepers.go
@@ -259,6 +259,7 @@ func (ak *AppKeepers) InitKeepers(
 		runtime.NewKVStoreService(keys[epochingtypes.StoreKey]),
 		bankKeeper,
 		stakingKeeper,
+		stakingkeeper.NewMsgServerImpl(stakingKeeper),
 		appparams.AccGov.String(),
 	)