+++ x/epoching/keeper/keeper.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	stktypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
 	corestoretypes "cosmossdk.io/core/store"
 	"cosmossdk.io/log"