+++ CHANGELOG.md
@@ -44,6 +44,7 @@ This PR contains a series of PRs on multi-staking support and BTC stakingintegra
 - [#391](https://github.com/babylonlabs-io/babylon/pull/391) Fix e2e `TestBTCRewardsDistribution` flunky
 check of rewards
 - [#419](https://github.com/babylonlabs-io/babylon/pull/419) Add new modules to swagger config
+- [#436](https://github.com/babylonlabs-io/babylon/pull/436) Remove staking msg server to handle at epoch
 
 ### Bug fixes